requirements.txt
flask
flask_cors
python-dotenv
pyjwt
bcrypt
pymongo